Quick Commit, update Inputs, Constructor
